% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{ExpProb}
\alias{ExpProb}
\title{Estimate Exposure Probabilities}
\usage{
ExpProb(formula, fit_dat, pred_dat)
}
\arguments{
\item{formula}{A formula for the exposure variable.}

\item{fit_dat}{A data used to fit the model.}

\item{pred_dat}{A data used to predict exposure probabilities.}
}
\value{
A vector of predicted exposure probabilities.
}
\description{
Description.....
This function estimates exposure probabilities by fitting a logistic regression model.
}
\details{
This function uses \code{speedglm} for fast computation and is based on data.frame()
which is the default format of speedglm(). And this function does not support
data.table().
}
\examples{
\dontrun{
result <- ExpProb(
   formula = A ~  X1 + X2 + X3,
   fit_dat = data_fit,
   pred_dat = data_pre
)
head(result)
}


}
