% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainfctAllT.R
\name{MainfctAllT}
\alias{MainfctAllT}
\title{Title ......}
\usage{
MainfctAllT(
  K,
  data,
  exp_fo,
  prin_fo,
  mu_fo,
  col_names,
  Y_type,
  n,
  S_col,
  A_col,
  Y_col,
  ind_col,
  time_col,
  id_col
)
}
\arguments{
\item{K}{Integer. Number of time points.}

\item{data}{A \code{data.table}-like dataset with variables including \code{id}, \code{time}, \code{Y}, \code{S}, \code{A}, \code{U11}, etc.}

\item{exp_fo}{A formula object for modeling exposure probabilities, passed to \code{ExpProb()}.}

\item{prin_fo}{A formula object for modeling principal scores, passed to \code{PrinPred()}.}

\item{mu_fo}{A formula object for modeling outcome means, passed to \code{MuPred()}.}

\item{col_names}{A character vector of selected column names used as covariates.}

\item{Y_type}{Integer. Type of data. 1 = continuous data, 2 = binary data.}

\item{n}{Integer. Number of regression coefficients, typically = #selected_covariates + intercept.}

\item{S_col}{The name of the survival status column.}

\item{A_col}{The name of the treatment status column.}

\item{Y_col}{The name of the variable outcome column.}

\item{ind_col}{The name of the indicator column.}

\item{time_col}{The name of time point column.}

\item{id_col}{The name of unique individual id column.}
}
\value{
beta. A 1 by n matrix of estimated bias-corrected coefficients.
}
\description{
Description.....
}
\details{
Computes time-varying coefficients using approach HTE,
handling both continuous and binary outcome types.

more details to be added.......
This function supports both linear and nonlinear link estimation via
Newton-Raphson with analytical Jacobian.
Besides, this function is based on the format of data.table for it help
save storage and improve calculation speed.
}
\examples{
\dontrun{
result <- MainfctAllT(
 i = 1, K = 4, data = BiSampleSmall,
 exp_fo = A ~  X1 + X2 + X3,
 prin_fo = S ~ (X1 + X2 + X3) * A,
 mu_fo = Y ~ (X1 + X2 + X3) * A + S - 1,
 Y_type = 1,
 col_names = c("X1","X2","X3"),
 n = 4,
 S_col = "S",
 A_col = "A",
 Y_col = "Y",
 ind_col = "ind",
 time_col = "time",
 id_col = "id"
)
result$beta
}

}
