% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{MuPred}
\alias{MuPred}
\title{Estimate Outcome Mean}
\usage{
MuPred(formula, fit_dat, pred_dat, trt, Y_type, A_col, S_col)
}
\arguments{
\item{formula}{A fomula for the outcome mean variable.}

\item{fit_dat}{A data used to fit the model.}

\item{pred_dat}{A data used to predict outcome mean.}

\item{trt}{Integer. Type of A. 1 = Treated, 0 = Not Treated.}

\item{Y_type}{Integer. Type of data. 1 = continuous data, 2 = binary data.}

\item{A_col}{The name of the treatment status column.}

\item{S_col}{The name of the survival status column.}
}
\value{
A vector of predicted outcome mean.
}
\description{
Description.....
This function estimates outcome mean.
}
\details{
This function supports two types of Y and is based on data.frame()
which is the default format of glm().
}
\examples{
\dontrun{
result <- MuPred(
   formula = A ~  X1 + X2 + X3,
   fit_dat = data_fit,
   pred_dat = data_pre,
   trt = 1,
   Y_type = 1,
   S_col = "S",
   A_col = "A"
)
head(result)
}


}
