% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{PrinPred}
\alias{PrinPred}
\title{Estimate Principal Score}
\usage{
PrinPred(formula, fit_dat, pred_dat, trt, ind_col, A_col)
}
\arguments{
\item{formula}{A fomula for the principal score variable.}

\item{fit_dat}{A data used to fit the model.}

\item{pred_dat}{A data used to predict principal score.}

\item{trt}{Integer. Type of A. 1 = Treated, 0 = Not Treated.}

\item{ind_col}{The name of the indicator column.}

\item{A_col}{The name of the treatment status column.}
}
\value{
A list with the following components
\describe{
\item{e_pred}{A vector of predicted principal score.}
\item{fit_ps_A}{The fitted logistic regression model object}
}
}
\description{
Description.....
This function estimates principal score by fitting a logistic regression model.
}
\details{
This function uses \code{speedglm} for fast computation and is based on data.frame()
which is the default format of speedglm(). And this function does not support
data.table().
}
\examples{
\dontrun{
result <- PrinPred(
   formula = A ~  X1 + X2 + X3,
   fit_dat = data_fit,
   pred_dat = data_pre,
   trt = 1,
   ind_col = "ind",
   A_col = "A"
)
head(result)
summary(result$fit_ps_A)
}


}
